Automating charts generation
As the 'cron' methodology does generate the charts in the specific
periods only, for less sophisticated machines (as Raspberry PI) this is
quite a waste of computing power.
Now the graphs are generated as per request.
To show the progress the 'calemsgraphs.sh' script from "ems_tools" was
divided into 5 separate scripts.